analysis/pyproject.toml
[project]
name = "baseball-lab-analysis"
version = "0.1.0"
dependencies = [
    "opencv-python",
    "mediapipe",
    "numpy",
    "yt-dlp",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "ultralytics>=8.4.33",
    "pyyaml>=6.0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/baseball_lab"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
